Remove TestCase.defer{SetUp, TestMethod, TearDown, RunCleanups}
Currently TestCase has the following functions that are presumably
public due to their naming: deferSetUp, deferTestMethod,
deferRunCleanups and deferTearDown. This raises a few problems. It makes
sense to remove the functions without deprecation period due to the
following reasons:

- these functions expose so much of the internal structure of TestCase
that it makes the code hard to change.
- there is no case where the functions would be useful as an extension
point. The only use of the functions that I found using Github code
search is the Buildbot project which overrides deferRunCleanups and
hooks even deeper into TestCase internals.
- the API is not "public ready" (i.e. the ignored parameter) - it seems
as if they were some kind of internal thing that simply didn't get the _
prefix.
- deprecation would be harder to implement than usual, due to the
functions being supposed to be overridden, not just called.

Given the above - small risk of breaking code for someone and relatively
high costs for going through the deprecation dance it makes sense to
remove this code